+++ package.json
@@ -32,6 +32,7 @@
     "passport-local": "^1.0.0"
   },
   "devDependencies": {
+    "eslint": "^4.14.0",
     "frisby": "^2.0.8",
     "jest": "^20.0.4",
     "jsdom": "^11.0.0",
