+++ src/init.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const accepts = require('accepts');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 const debug = require('debug')('skycap');
@@ -9,7 +8,8 @@ const session = require('express-session');
 const flash = require('connect-flash');
 
 const { config, mergeConfig } = require('./config');
-const SkycapAuth = require('./auth');
+const { requireUser } = require('./middleware');
+const SkycapUserSDK = require('./sdk/user');
 const SkycapUser = require('./user');
 
 // Templates
@@ -34,7 +34,7 @@ function mount(app, adapterClass, options) {
     throw new Error('An auth adapter must be set - please specify adapterClass');
   }
 
-  let authAdapter = new SkycapAuth(new adapterClass())
+  let userSdk = new SkycapUserSDK(new adapterClass())
 
   // Setup database sessions
   app.use(adapterClass.useSession(session, config));
@@ -45,7 +45,7 @@ function mount(app, adapterClass, options) {
   });
 
   passport.deserializeUser(function(email, done) {
-    authAdapter.findByEmail(email)
+    userSdk.findByEmail(email)
       .then((user) => {
         let userObj = new SkycapUser(user);
 
@@ -67,7 +67,7 @@ function mount(app, adapterClass, options) {
   passport.use(new LocalStrategy({
     usernameField: 'email'
   }, function(email, password, done) {
-    authAdapter.findByEmailAndPassword(email, password)
+    userSdk.findByEmailAndPassword(email, password)
       .then((user) => {
         done(null, user);
       })
@@ -109,7 +109,7 @@ function mount(app, adapterClass, options) {
     let profileData = null; // @TODO: Make this customizeable
 
     // Register new user
-    authAdapter.register(email, password, profileData)
+    userSdk.register(email, password, profileData)
       .then((user) => {
         // Log user in after successful registration
         req.login(user, function(err) {
@@ -146,149 +146,5 @@ function mount(app, adapterClass, options) {
   return (req, res, next) => next();
 }
 
-/**
- * Create new user object with given data
- *
- * @param data object
- * @return SkycapUser
- */
-function createUserObject(data) {
-  console.log('[skycap/init] createUserObject', data);
-  return new SkycapUser(data);
-}
 
-/**
- * Format for API error
- */
-function apiResponse(json, type) {
-  let output = {
-    data: {
-      type,
-      id: json.id,
-      attributes: json,
-    }
-  };
-
-  delete output.data.attributes.id;
-
-  return output;
-}
-
-/**
- * create API error response
- */
-function apiErrorResponse(err, statusCode = 500) {
-  let json = {
-    errors: [
-      {
-        status: String(statusCode),
-        title: err.message || err,
-      },
-    ],
-  };
-
-  return json;
-}
-
-/**
- * Use passport auth with redirects
- */
-function passportRedirectAuth(req, res, next) {
-  passport.authenticate('local', function(err, user, info) {
-    let defaultMessage = config.errors.user_bad_auth;
-
-    if (err) {
-      req.flash('skycap_errors', err.message);
-      return next(err);
-    }
-
-    if (!user) {
-      req.flash('skycap_errors', info.message || defaultMessage);
-      return res.redirect(config.routes.user.login);
-    }
-
-
-    req.logIn(user, function(err) {
-      if (err) {
-        req.flash('skycap_errors', err.message);
-        return next(err);
-      }
-
-      return res.redirect(config.redirects.loginSuccess);
-    });
-  })(req, res, next);
-}
-
-/**
- * Use passport auth with custom response for API
- */
-function passportApiAuth(req, res, next) {
-  passport.authenticate('local', function(err, user, info) {
-    let unauthorizedStatus = 401;
-    let accept = accepts(req);
-    let isJson = accept.type(['html', 'json']) === 'json';
-
-    if (err) {
-      return res.status(unauthorizedStatus).json(apiErrorResponse(err, unauthorizedStatus));
-    }
-
-    if (!user) {
-      return res.status(unauthorizedStatus).json(apiErrorResponse('User not logged in', unauthorizedStatus));
-    }
-
-
-    req.logIn(user, function(err) {
-      if (err) {
-        return next(err);
-      }
-
-      let userObj = createUserObject(user);
-
-      return res.status(200).json(apiResponse(userObj.toJSON(), 'user'));
-    });
-  })(req, res, next);
-}
-
-/**
- * Main method of enforcing user logins and sessions
- */
-function requireAdmin() {
-  return function (req, res, next) {
-    let accept = accepts(req);
-    let isJson = accept.type(['html', 'json']) === 'json';
-
-    if (req.isAuthenticated() && req.user.isAdmin()) {
-      return next();
-    }
-
-    if (isJson) {
-      return passportApiAuth(req, res, next);
-    } else {
-      return passportRedirectAuth(req, res, next);
-    }
-  };
-}
-
-/**
- * Main method of enforcing user logins and sessions
- */
-function requireUser() {
-  return function (req, res, next) {
-    let accept = accepts(req);
-    let isJson = accept.type(['html', 'json']) === 'json';
-
-    if (req.isAuthenticated()) {
-      debug('User already authenticated as =', req.user);
-      return next();
-    }
-
-
-    if (isJson) {
-      return passportApiAuth(req, res, next);
-    } else {
-      return passportRedirectAuth(req, res, next);
-    }
-  };
-}
-
-module.exports = { requireAdmin, requireUser, mount };
+module.exports = { mount };
